requirements/base.txt
django==1.11
Pillow
django-bootstrap3
django-allauth
mysqlclient
django-simple-history==1.9.0
django-debug-toolbar
pytils==0.3
sorl-thumbnail==12.3